requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
cryptography==3.3.1
defusedxml==0.6.0
idna==2.10
jira==2.0.0
oauthlib==3.1.0
pbr==5.5.1
prettytable==2.0.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.0.1
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rich==10.9.0
six==1.15.0
urllib3==1.26.2
wcwidth==0.2.5